requirements.txt
boto3
flask
six
pillow
json_logging
tensorflow_gpu
tensorflow_hub